+++ tests/toisto/persistence/test_progress.py
@@ -4,10 +4,12 @@ import unittest
 from argparse import ArgumentParser
 from unittest.mock import MagicMock, Mock, patch
 
+from toisto.metadata import get_progress_filepath
+from toisto.model.language import Language
 from toisto.model.quiz.progress import Progress
 from toisto.model.quiz.retention import Retention
 from toisto.model.quiz.topic import Topics
-from toisto.persistence.progress import PROGRESS_JSON, load_progress, save_progress
+from toisto.persistence.progress import load_progress, save_progress
 
 
 class ProgressPersistenceTest(unittest.TestCase):
@@ -17,7 +19,7 @@ class ProgressPersistenceTest(unittest.TestCase):
     @patch("pathlib.Path.open", MagicMock())
     def test_load_non_existing_progress(self):
         """Test that the default value is returned when the progress cannot be loaded."""
-        self.assertEqual({}, load_progress(Topics(set()), ArgumentParser()).as_dict())
+        self.assertEqual({}, load_progress(Topics(set()), "fi", ArgumentParser()).as_dict())
 
     @patch("pathlib.Path.exists", Mock(return_value=True))
     @patch("sys.stderr.write")
@@ -25,22 +27,27 @@ class ProgressPersistenceTest(unittest.TestCase):
     def test_load_invalid_progress(self, path_open: Mock, stderr_write: Mock):
         """Test that the the program exists if the progress cannot be loaded."""
         path_open.return_value.__enter__.return_value.read.return_value = ""
-        self.assertRaises(SystemExit, load_progress, Topics(set()), ArgumentParser())
-        self.assertIn(f"cannot parse the progress information in {PROGRESS_JSON}", stderr_write.call_args_list[1][0][0])
+        language = Language("en")
+        self.assertRaises(SystemExit, load_progress, Topics(set()), language, ArgumentParser())
+        progress_file = get_progress_filepath(language)
+        self.assertIn(f"cannot parse the progress information in {progress_file}", stderr_write.call_args_list[1][0][0])
 
     @patch("pathlib.Path.exists", Mock(return_value=True))
     @patch("pathlib.Path.open")
     def test_load_existing_progress(self, path_open: Mock):
         """Test that the progress can be loaded."""
         path_open.return_value.__enter__.return_value.read.return_value = '{"quiz": {}}'
-        self.assertEqual(dict(quiz=Retention().as_dict()), load_progress(Topics(set()), ArgumentParser()).as_dict())
+        self.assertEqual(
+            dict(quiz=Retention().as_dict()),
+            load_progress(Topics(set()), Language("nl"), ArgumentParser()).as_dict(),
+        )
 
     @patch("pathlib.Path.open")
     @patch("json.dump")
     def test_save_empty_progress(self, dump: Mock, path_open: Mock):
         """Test that the progress can be saved."""
         path_open.return_value.__enter__.return_value = json_file = MagicMock()
-        save_progress(Progress({}, Topics(set())))
+        save_progress(Progress({}, Topics(set()), Language("fi")))
         dump.assert_called_once_with({}, json_file)
 
     @patch("pathlib.Path.open")
@@ -48,7 +55,7 @@ class ProgressPersistenceTest(unittest.TestCase):
     def test_save_incorrect_only_progress(self, dump: Mock, path_open: Mock):
         """Test that the progress can be saved."""
         path_open.return_value.__enter__.return_value = json_file = MagicMock()
-        save_progress(Progress(dict(quiz={}), Topics(set())))
+        save_progress(Progress(dict(quiz={}), Topics(set()), Language("fi")))
         dump.assert_called_once_with(dict(quiz={}), json_file)
 
     @patch("pathlib.Path.open")
@@ -56,5 +63,5 @@ class ProgressPersistenceTest(unittest.TestCase):
     def test_save_progress(self, dump: Mock, path_open: Mock):
         """Test that the progress can be saved."""
         path_open.return_value.__enter__.return_value = json_file = MagicMock()
-        save_progress(Progress(dict(quiz=dict(skip_until="3000-01-01")), Topics(set())))
+        save_progress(Progress(dict(quiz=dict(skip_until="3000-01-01")), Topics(set()), Language("fi")))
         dump.assert_called_once_with(dict(quiz=dict(skip_until="3000-01-01T00:00:00")), json_file)